Moved prog gen pyapi to metal
