+++ rust/pyapi/src/prog_gen/mod.rs
@@ -1,94 +1,33 @@
 //! Implements Python bindings for program generation data structures and functions
 
-mod condition;
-pub mod flow_options;
-pub mod group;
-pub mod interface;
-mod pattern_group;
-mod resources;
-mod test;
-mod test_invocation;
-
-pub use condition::Condition;
-pub use group::Group;
 use origen::core::tester::TesterSource;
-use origen_metal::prog_gen::{flow_api, FlowCondition, Model, ParamType, ParamValue};
-use origen_metal::FLOW;
+use origen_metal::prog_gen::Model;
 use origen_metal::prog_gen::SupportedTester;
-use origen::{Error, Result};
-pub use pattern_group::PatternGroup;
 use pyo3::prelude::*;
-use pyo3::types::PyAny;
 use pyo3::wrap_pyfunction;
-use resources::Resources;
 use std::collections::HashMap;
 use std::thread;
-pub use test::Test;
-pub use test_invocation::TestInvocation;
+use regex::Regex;
+use std::path::PathBuf;
 
 pub fn define(py: Python, m: &PyModule) -> PyResult<()> {
     let subm = PyModule::new(py, "prog_gen")?;
-    subm.add_wrapped(wrap_pyfunction!(start_new_flow))?;
-    subm.add_wrapped(wrap_pyfunction!(end_flow))?;
     subm.add_wrapped(wrap_pyfunction!(render))?;
+    subm.add_wrapped(wrap_pyfunction!(resolve_file_reference))?;
     m.add_submodule(subm)?;
     Ok(())
 }
 
 #[pyfunction]
-fn start_new_flow(
-    name: &str,
-    sub_flow: Option<bool>,
-    bypass_sub_flows: Option<bool>,
-    add_flow_enable: Option<&str>,
-) -> PyResult<Vec<usize>> {
-    let sub_flow = match sub_flow {
-        None => false,
-        Some(x) => x,
-    };
-    let mut refs = vec![];
-    if sub_flow {
-        refs.push(flow_api::start_sub_flow(name, None, None)?);
-    } else {
-        FLOW.start(name)?;
-        refs.push(0);
-        if let Some(bypass) = bypass_sub_flows {
-            if bypass {
-                refs.push(flow_api::start_bypass_sub_flows(None)?);
-            }
+fn resolve_file_reference(path: &str) -> PyResult<String> {
+    let file = origen::with_current_job(|job| {
+        let mut pt = PathBuf::from(".");
+        for p in Regex::new(r"(\\|/)").unwrap().split(path) {
+            pt.push(p);
         }
-        if let Some(enable) = add_flow_enable {
-            let flag = format!("{}_enable", name);
-            refs.push(flow_api::start_condition(
-                FlowCondition::IfEnable(vec![flag.clone()]),
-                None,
-            )?);
-            if enable.to_lowercase() == "enabled" {
-                flow_api::set_default_flag_state(flag, true, None)?;
-            } else if enable.to_lowercase() == "disabled" {
-                flow_api::set_default_flag_state(flag, false, None)?;
-            } else {
-                return Err(PyErr::from(Error::new(&format!(
-                    "The add_flow_enable argument must be either None (default), \"enabled\" or \"disabled\", got '{}'",
-                    enable
-                ))));
-            }
-        }
-    }
-    refs.reverse();
-    Ok(refs)
-}
-
-#[pyfunction]
-fn end_flow(ref_ids: Vec<usize>) -> PyResult<()> {
-    for ref_id in ref_ids {
-        if ref_id == 0 {
-            FLOW.end()?;
-        } else {
-            flow_api::end_block(ref_id)?;
-        }
-    }
-    Ok(())
+        job.resolve_file_reference(&pt, Some(vec!["py"]))
+    })?;
+    Ok(file.to_str().unwrap().to_string())
 }
 
 // Called automatically by Origen once all test program source files have been executed
@@ -151,93 +90,3 @@ fn render(py: Python) -> PyResult<Vec<String>> {
     })
 }
 
-pub fn to_param_value(value: &PyAny) -> Result<Option<ParamValue>> {
-    Ok(if let Ok(v) = value.extract::<bool>() {
-        Some(ParamValue::Bool(v))
-    } else if let Ok(v) = value.extract::<u64>() {
-        Some(ParamValue::UInt(v))
-    } else if let Ok(v) = value.extract::<i64>() {
-        Some(ParamValue::Int(v))
-    } else if let Ok(v) = value.extract::<f64>() {
-        Some(ParamValue::Float(v))
-    } else if let Ok(v) = value.extract::<String>() {
-        Some(ParamValue::String(v))
-    } else if let Ok(None) = value.extract::<Option<String>>() {
-        None
-    } else {
-        Some(ParamValue::Any(format!("{}", value.str()?)))
-    })
-}
-
-#[allow(dead_code)] // Could be used in future
-pub fn to_param_value_with_type(ptype: &ParamType, value: &PyAny) -> Result<ParamValue> {
-    match ptype {
-        ParamType::Bool => {
-            if let Ok(v) = value.extract::<bool>() {
-                Ok(ParamValue::Bool(v))
-            } else {
-                bail!("Illegal value, expected a Boolean, got: '{}'", value)
-            }
-        }
-        ParamType::Int => {
-            if let Ok(v) = value.extract::<i64>() {
-                Ok(ParamValue::Int(v))
-            } else {
-                bail!("Illegal value, expected an Integer, got: '{}'", value)
-            }
-        }
-        ParamType::UInt => {
-            if let Ok(v) = value.extract::<u64>() {
-                Ok(ParamValue::UInt(v))
-            } else {
-                bail!(
-                    "Illegal value, expected an Unsigned Integer, got: '{}'",
-                    value
-                )
-            }
-        }
-        ParamType::Float => {
-            if let Ok(v) = value.extract::<f64>() {
-                Ok(ParamValue::Float(v))
-            } else {
-                bail!("Illegal value, expected a Float, got: '{}'", value)
-            }
-        }
-        ParamType::Current => {
-            if let Ok(v) = value.extract::<f64>() {
-                Ok(ParamValue::Current(v))
-            } else {
-                bail!("Illegal value, expected a Float, got: '{}'", value)
-            }
-        }
-        ParamType::Voltage => {
-            if let Ok(v) = value.extract::<f64>() {
-                Ok(ParamValue::Voltage(v))
-            } else {
-                bail!("Illegal value, expected a Float, got: '{}'", value)
-            }
-        }
-        ParamType::Time => {
-            if let Ok(v) = value.extract::<f64>() {
-                Ok(ParamValue::Time(v))
-            } else {
-                bail!("Illegal value, expected a Float, got: '{}'", value)
-            }
-        }
-        ParamType::Frequency => {
-            if let Ok(v) = value.extract::<f64>() {
-                Ok(ParamValue::Frequency(v))
-            } else {
-                bail!("Illegal value, expected a Float, got: '{}'", value)
-            }
-        }
-        ParamType::String => {
-            if let Ok(v) = value.extract::<String>() {
-                Ok(ParamValue::String(v))
-            } else {
-                bail!("Illegal value, expected a String, got: '{}'", value)
-            }
-        }
-        ParamType::Any => Ok(ParamValue::Any(format!("{}", value.str()?))),
-    }
-}
