+++ rust/pyapi_metal/src/prog_gen/condition.rs
@@ -1,4 +1,4 @@
-use crate::utility::caller::src_caller_meta;
+use super::src_caller_meta;
 use origen_metal::prog_gen::{flow_api, FlowCondition};
 use pyo3::prelude::*;
 use pyo3::types::{PyAny, PyType};
