+++ rust/pyapi_metal/src/prog_gen/flow_options.rs
@@ -1,8 +1,8 @@
 //! Functions concerned with extracting known values from kwargs
 
-use crate::utility::caller::src_caller_meta;
+use super::src_caller_meta;
 use origen_metal::prog_gen::{flow_api, FlowCondition, FlowID};
-use origen::Result;
+use origen_metal::Result;
 use pyo3::types::PyDict;
 
 pub fn is_flow_option(key: &str) -> bool {
