+++ rust/pyapi_metal/src/prog_gen/group.rs
