+++ rust/pyapi_metal/src/prog_gen/test.rs
@@ -1,8 +1,8 @@
 use super::to_param_value;
 use crate::prog_gen::flow_options;
-use crate::utility::caller::src_caller_meta;
+use super::src_caller_meta;
 use origen_metal::prog_gen::{flow_api, Limit, LimitSelector, ParamValue, SupportedTester};
-use origen::Result;
+use origen_metal::Result;
 use pyo3::prelude::*;
 use pyo3::types::PyDict;
