+++ rust/pyapi_metal/src/prog_gen/test_invocation.rs
@@ -1,9 +1,9 @@
 use super::to_param_value;
 use super::Test;
 use crate::prog_gen::flow_options;
-use crate::utility::caller::src_caller_meta;
+use super::src_caller_meta;
 use origen_metal::prog_gen::{flow_api, Limit, LimitSelector, ParamValue, SupportedTester};
-use origen::Result;
+use origen_metal::Result;
 use pyo3::exceptions::PyAttributeError;
 use pyo3::prelude::*;
 use pyo3::types::PyDict;
